port osal_FreeRTOS, add some include to be able to mock & link change osal_none's queue_t member from uint16_t to uint8_t

+++ tinyusb/osal/osal_none.h
@@ -142,10 +142,10 @@ static inline  tusb_error_t osal_semaphore_post(osal_semaphore_handle_t const se
 //--------------------------------------------------------------------+
 typedef struct{
            uint32_t * const buffer     ; ///< buffer pointer
-           uint16_t const depth        ; ///< buffer size
-  volatile uint16_t count          ; ///< bytes in fifo
-  volatile uint16_t wr_idx       ; ///< write pointer
-  volatile uint16_t rd_idx       ; ///< read pointer
+           uint8_t const depth        ; ///< buffer size
+  volatile uint8_t count          ; ///< bytes in fifo
+  volatile uint8_t wr_idx       ; ///< write pointer
+  volatile uint8_t rd_idx       ; ///< read pointer
 } osal_queue_t;
 
 typedef osal_queue_t * osal_queue_handle_t;
